Etherman optionally readonly. (#1321)
Closes:  https://github.com/0xPolygonHermez/zkevm-node/issues/1227

This PR implements the option to set Etherman to be read only (a.k.a. without an account). This is useful for nodes that don't send transactions.

The behavior is controlled by the new ReadOnly field in the Etherman config (false by default).

Methods accessing the account are now safeguarded by a check on the readOnly field in the Etherman struct.

Bonus: added the Etherman test cases in config/config_test.go

Includes:

* No need to supply RPC with account (#1337)

Co-authored-by: Konrad Iturbe <KonradIT@users.noreply.github.com>