fix: wait to be synced after sending final proof
This commit fixes a bug that was causing the aggregato to start
generating a duplicated proof.
The problem was that after sending the final proof to L1, the aggregator
was not waiting for it to be seen by the synchronizer. Instead it was
checking immediately for batches to prove. Since the synchronizer didn't
update the verified batches yet, the aggregator was picking up again the
first virtual batch which followed the previous verified batch.
This commit adds a signal emitted after a final proof is sent. The next
aggregator iteration will get the signal and wait for the synchronizer
to sync.

Signed-off-by: Paolo Grisoli <paolo_g84@icloud.com>