fix: add check on mocked LER and SR
This commit adds a check (similar to the one in the old aggregator) to
check if we are using a mocked prover. In this case the `NewStateRoot`
and `NewLocalExitRoot` values returned by the mock prover are not valid
and need to be replaced with the ones returned by the executor.