feat: setup log by environment (dev, prod)
This commit removes the `encoding` value from the `Log` configuration,
in favor of an `environment` value. This new value can be set to
`production` or `development` and it will set up the logs in a suitable
format (i.e. for `production` logs in json, time in epoch, for
`development` colored log levels, human readable timestamps, ...).
Credits to @tclemos for pointing to the underlying zap functions.

Signed-off-by: Paolo Grisoli <paolo_g84@icloud.com>