[Aggregator] Send pub address on final proof request (#1455)
* fix: update aggregator proto to send pub address

This commit applies the changes introduced to the aggregator proto file.
The changes are regarding the GetGenFinalProofRequest message that now
needs to send the aggregator public address.

* chore: add goreleaser file

* Get NewLocalExitRoot and NewStateRoot from last batch in the aggregated proof

* feat: add prover IP address

This commit adds the prover IP address to the Prover struct. The address
is retrieved from the grpc connection and logged alogside the prover ID.

* fix: remove ShortID method from Prover

This commit removes the ShortID method from the Prover. This because the
prover ID can be a fixed arbitrary string and not only a UUID.

* fix accinputhash (#1458)

* refac: send final proof in FinalProofInputs

This commit modifies the FinalProofInputs struct to hold also the final
proof. The VerifyBatches signature has been modified accordingly.

Co-authored-by: zkronos73 <zkronos73@gmail.com>
Co-authored-by: Alonso Rodriguez <ARR552@users.noreply.github.com>