feat: cleanup proofs on verification
This commit adds a mechanism to cleanup the proof table after a final
proof has been verified.
The `isSynced` method has been modified to accept an optional `batchNum`
argument. If this is provided, then `isSynced` will make sure that the
state is synced up to that batch number.
This allows the Aggregator to be sure to delete all the records from the
`proof` table up to that batch number.

Signed-off-by: Paolo Grisoli <paolo_g84@icloud.com>