Fix encoding bug in AddForcedBatch
The forced_batch table expects the raw_txs_data column to be a
string (specifically a hex encoding of the raw transaction bytes).
The corresponding GetForcedBatch function uses hex.DecodeString to
turn the value of this column back into raw bytes. However,
AddForcedBatch is missing a hex.EncodeToString, so it is actually
interpreting the raw transaction bytes as an ASCII representation
of a hex-encoded string.