Add UTs to the aggregator package (#1583)
Closes #1578 .

This PR adds unit tests to the `aggregator` package.
It also introduces a small change in the `sendFinalProof` logic: after the final proof is successfully verified, instead of deleting form the `proof` table only the batches related to that final proof, now it deletes **all** the batches **up to** the latest verified (using the new `state` `CleanupGeneratedProofs` method). In this way we avoid to have stale batches in the `proof` table for whatever reason (network error, crashes, ...).

Logs have also been polished to take advantage of the `WithFields` feature as much as possible.