[Aggregator] Add recovery mechanism for stuck proofs (#1603)
Closes #1516 .

### Background

There may be chances in which a proof under generation by a prover end up being lost (network failure, db failure, ...).
These proofs stays forever in our `proof` table because they looked to be reserved by a prover, thus the aggregator does not send them to other provers.
For this reason we needed a mechanism to clear those stuck proofs from the table.

### Proposal

This PR  introduces a new method in the `Aggregator` called `cleanupLockedProofs` that continuously checks for locked proofs in the table and if any it clears them out.
The frequency of this check and the threshold from which a proof is considered stuck are configurable with `TimeCleanupLockedProofs` and `GeneratingProofCleanupThreshold` respectively.

The `Generating` (`bool`) field has been refactored to `GeneratingSince` (`*time.Time`) to allow the time comparison against the `GeneratingProofCleanupThreshold` config value. If the field is `nil` it means that the proof is not currently being generating by a prover, otherwise we have the time in which the generation started.

Additional two fields have been added to the `Proof` struct: `CreatedAt` and `UpdatedAt`. For now they don't have a specific usecase, but the may come useful in debugging and troubleshooting.

Other stuff included:

- added UT in `state/pgstatestorage_test.go` for `CleanupLockedProofs`
- fixed `batchsender` script
- added Aggregator section in `config/config_test.go`
- Makefile `sail` and `resail` commands rebranded `ship` and `reship`
- cleaned up `go.sum` by running `go mod tidy -compat=1.17` (:warning: if we need to keep compatibility this can be reverted and run `go mod tidy -go=1.16 && go mod tidy -go=1.17` instead)