chore: add proofId and batches to prover error log
This commit adds the `proofId` and `batches` (or `batch`) fields to the
error messages returned in case of prover errors happening at proof
generation time. This allows to print these additional information in
the logs from the outer scope (in the `Channel` method), where `proofId`
and `batches` are not available.