fix: add formatting in log template for the ST
This commit adds an additional formatting to the error message template,
to hold the additional argument that is appended for the stacktrace.
This applies to `Errorf` and `Fatalf`.
Also the first argument of the 'w' functions (`Infow`, `Debugw`, ...)
has been renamed `msg`, the previous `template` name was misleading
because actually it is not treated as a template.